+++ lib/circular_countdown.dart
@@ -1,3 +1,4 @@
 library circular_countdown;
 
 export 'src/circular_countdown.dart';
+export 'src/time_circular_countdown.dart';